requirements.txt
torch==2.1.2
torchvision
transformers
scikit-learn
Pillow
lightning
fastapi
uvicorn
